# copilot-studio-agent/solution/connectors/hr-api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contoso HR API (Copilot Studio Custom Connector)",
    "version": "1.1.0",
    "description": "Custom Connector for the Contoso HR API exposed through APIM. Field names and response shapes match the FastAPI backend exactly so Power Automate flow expressions like outputs('Create_leave_request')['managerId'] resolve correctly."
  },
  "host": "${HR_API_HOST}",
  "basePath": "/hr",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "apim_subscription_key": {
      "type": "apiKey",
      "in": "header",
      "name": "Ocp-Apim-Subscription-Key"
    }
  },
  "security": [
    {
      "apim_subscription_key": []
    }
  ],
  "paths": {
    "/employees/{id}": {
      "get": {
        "operationId": "getEmployee",
        "summary": "Get employee",
        "description": "Get an employee by id or email.",
        "x-ms-summary": "Get employee",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Employee id or email",
            "x-ms-visibility": "important"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "404": {
            "description": "Employee not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/employees/byName": {
      "get": {
        "operationId": "findEmployeesByName",
        "summary": "Find employees by name",
        "description": "Returns employees whose display name contains the supplied substring (case-insensitive). Names are not unique, so the response is always an array.",
        "x-ms-summary": "Find employees by name",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Full or partial name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "matches": {
                  "type": "array",
                  "x-ms-summary": "Matches",
                  "items": {
                    "$ref": "#/definitions/EmployeeRef"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees/byPersona": {
      "get": {
        "operationId": "findEmployeeByPersona",
        "summary": "Find employee by persona",
        "description": "Returns the first employee matching the supplied persona (e.g. buddy, manager, hr-partner, it).",
        "x-ms-summary": "Find employee by persona",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "persona",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Persona"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EmployeeRef"
            }
          },
          "404": {
            "description": "No employee with that persona",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/leave/request": {
      "post": {
        "operationId": "requestLeave",
        "summary": "Request leave",
        "description": "Create a leave request. Returns the full request including the resolved manager id and computed day count, so the calling flow can drive an Approvals connector step without a second lookup.",
        "x-ms-summary": "Request leave",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "type": "string",
            "x-ms-summary": "Idempotency key",
            "x-ms-visibility": "advanced"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "employeeId",
                "startDate",
                "endDate"
              ],
              "properties": {
                "employeeId": {
                  "type": "string",
                  "x-ms-summary": "Employee id"
                },
                "startDate": {
                  "type": "string",
                  "x-ms-summary": "Start date (yyyy-mm-dd)"
                },
                "endDate": {
                  "type": "string",
                  "x-ms-summary": "End date (yyyy-mm-dd)"
                },
                "type": {
                  "type": "string",
                  "x-ms-summary": "Leave type",
                  "default": "vacation",
                  "enum": [
                    "vacation",
                    "sick",
                    "personal"
                  ]
                },
                "note": {
                  "type": "string",
                  "x-ms-summary": "Note",
                  "x-ms-visibility": "advanced"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/LeaveRequest"
            }
          },
          "200": {
            "description": "OK (idempotent replay)",
            "schema": {
              "$ref": "#/definitions/LeaveRequest"
            }
          },
          "400": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          },
          "404": {
            "description": "Employee not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/leave/{id}": {
      "get": {
        "operationId": "getLeave",
        "summary": "Get leave request",
        "x-ms-summary": "Get leave request",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Leave request id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LeaveRequest"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/leave/{id}/approve": {
      "post": {
        "operationId": "approveLeave",
        "summary": "Approve leave",
        "description": "Idempotent \u2014 approving an already-approved request is a no-op.",
        "x-ms-summary": "Approve leave",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Leave request id"
          },
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Decision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LeaveRequest"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          },
          "409": {
            "description": "Conflicting state",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/leave/{id}/reject": {
      "post": {
        "operationId": "rejectLeave",
        "summary": "Reject leave",
        "description": "Idempotent \u2014 rejecting an already-rejected request is a no-op.",
        "x-ms-summary": "Reject leave",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Leave request id"
          },
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Decision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LeaveRequest"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          },
          "409": {
            "description": "Conflicting state",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/onboarding/start": {
      "post": {
        "operationId": "startOnboarding",
        "summary": "Start onboarding",
        "x-ms-summary": "Start onboarding",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "newHireId",
                "startDate",
                "managerId"
              ],
              "properties": {
                "newHireId": {
                  "type": "string",
                  "x-ms-summary": "New hire id"
                },
                "startDate": {
                  "type": "string",
                  "x-ms-summary": "Start date"
                },
                "managerId": {
                  "type": "string",
                  "x-ms-summary": "Manager id"
                },
                "buddyId": {
                  "type": "string",
                  "x-ms-summary": "Buddy id"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/OnboardingPlan"
            }
          },
          "404": {
            "description": "Employee not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/onboarding/{planId}": {
      "get": {
        "operationId": "getOnboarding",
        "summary": "Get onboarding plan",
        "x-ms-summary": "Get onboarding plan",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Plan id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OnboardingPlan"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
    "/onboarding/{planId}/task/{taskId}/status": {
      "post": {
        "operationId": "setOnboardingTaskStatus",
        "summary": "Set onboarding task status",
        "x-ms-summary": "Set onboarding task status",
        "x-ms-visibility": "advanced",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Plan id"
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-ms-summary": "Task id"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "status"
              ],
              "properties": {
                "status": {
                  "type": "string",
                  "x-ms-summary": "Status",
                  "enum": [
                    "not-started",
                    "in-progress",
                    "done",
                    "overdue"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean"
                },
                "plan": {
                  "$ref": "#/definitions/OnboardingPlan"
                }
              }
            }
          },
          "404": {
            "description": "Plan or task not found",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Problem": {
      "type": "object",
      "description": "RFC 7807 problem details.",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      }
    },
...
    "Employee": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-ms-summary": "Employee id"
        },
        "displayName": {
          "type": "string",
          "x-ms-summary": "Display name"
        },
        "email": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "department": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "managerId": {
          "type": "string"
        },
        "persona": {
          "type": "string"
        }
      }
    },
    "EmployeeRef": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
  }
}